# pos_be/openapi.yaml
openapi: 3.0.3
info:
  title: GoFrugal POS API
  version: 1.0.0
  description: API documentation for GoFrugal POS System
paths:
...
  /api/sales/payments/:
    get:
      operationId: sales_payments_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - sales
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPaymentList'
          description: ''
    post:
      operationId: sales_payments_create
      tags:
      - sales
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Payment'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Payment'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: ''
...
components:
  schemas:
...
    PaginatedPaymentList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        bill:
          type: integer
        bill_number:
          type: string
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        payment_method:
          $ref: '#/components/schemas/PaymentPaymentMethodEnum'
        transaction_id:
          type: string
          nullable: true
          maxLength: 100
        payment_details:
          type: object
          additionalProperties: {}
          nullable: true
        status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        created_by:
          type: integer
          readOnly: true
          nullable: true
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - bill
      - bill_number
      - created_at
      - created_by
      - created_by_name
      - id
      - payment_method
      - updated_at
...
    PaymentPaymentMethodEnum:
      enum:
      - cash
      - card
      - upi
      - wallet
      - sodexo
      - store_credit
      - exchange
      type: string
      description: |-
        * `cash` - Cash
        * `card` - Card
        * `upi` - UPI
        * `wallet` - Wallet
        * `sodexo` - Sodexo
        * `store_credit` - Store Credit
        * `exchange` - Exchange
...
    PaymentStatusEnum:
      enum:
      - pending
      - completed
      - failed
      - refunded
      type: string
      description: |-
        * `pending` - Pending
        * `completed` - Completed
        * `failed` - Failed
        * `refunded` - Refunded
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT